Added ChiSquareAdapter. Changed degrees of freedom to int in CumulativeDistributionFunctionProvider interface.

+++ moa/src/main/java/moa/clusterers/outliers/SPLL/CumulativeDistributionFunctionProvider.java
@@ -2,6 +2,6 @@ package moa.clusterers.outliers.SPLL;
 
 public interface CumulativeDistributionFunctionProvider {
 	
-	double cumulativeProbability(double x, double df);
+	double cumulativeProbability(double x, int df);
 	
 }
